Updating with Ryan's suggested fixes
All instances of "floor plan" replaced with "floorplan". 
Renamed "Installing software" to "Getting Started" to account for non-software prerequisites. 
Included note that a dev token is needed for submission in the Getting Started section. 
Clarified that example image file types are examples - have purposefully used these are they are lossless formats and skipped .jpg as it is not. 
Have updated first instance of "latlongs" to "latitudes and longitudes" for clarity. 
Have updated "coordinate Q" to "coordinate (a, b)" for clarity. 
Have added a note explaining why we use a line-type shapefile rather than QGIS's "scratch layer” feature.